requirements.txt
bcrypt==3.2.0
cffi==1.14.3
click==7.1.2
cryptography==3.1.1
ecdsa==0.14.1
fastapi==0.61.1
h11==0.10.0
passlib==1.7.2
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycparser==2.20
pydantic==1.6.1
python-jose==3.2.0
python-multipart==0.0.5
rsa==4.6
six==1.15.0
sqlalchemy==1.3.19
starlette==0.13.6
uvicorn==0.12.1
